+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [0.2.4] - 2026-08-19
+
+- compaction aligned with upstream codex
+- $manifest brought back
+
 ## [0.2.3] - 2026-08-18
 
 - update notice says `Run bcodex update` again <- fixed